+++ src/list.jsx
@@ -1,13 +1,17 @@
 import React from 'react';
 
+import OneLine from './oneLine';
+
 export default function List({ todoList, onClickComplete }) {
   return (
     <ol>
-      {todoList.map((i, index) => (
-        <li key={i}>
-          {i}
-          <button type="button" onClick={() => onClickComplete(index)}>완료</button>
-        </li>
+      {todoList.map((element, index) => (
+        <OneLine
+          key={String(index)}
+          index={index}
+          element={element}
+          onClickComplete={onClickComplete}
+        />
       ))}
     </ol>
   );
